robot-api/requirements.txt
anyio==3.7.0
autopep8==2.0.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
codecov==2.1.13
colorama==0.4.6
coverage==7.2.6
deprecation==2.1.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
grpcio==1.54.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.25.0
orjson==3.8.14
packaging==23.1
pandas==2.0.2
protobuf==4.23.2
psycopg2==2.9.6
pycodestyle==2.10.0
pydantic==1.10.8
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
ta==0.10.2
tinkoff==0.1.1
tinkoff-investments==0.2.0b57
tomli==2.0.1
typing_extensions==4.6.2
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
